Generate and capture a full build log
It is placed into a `logs/` directory in the package dataset, with
the naming scheme:

```
<sourcepackage-basename>_<timestamp>_<cpuarch>.txt

```

Log files are not put into Git directly (by default), because they can
be large and there can be a lot of them.

Filenames are intended to be easily sortable.

Logs identify important steps in the build process and annotate them
with timestamps.

The build setup now uses the `chronic` helper to avoid including base
environment update information. These are not specific to a particular
package build and are only useful in case things fail.

Fixes datalad-debian#36