Add proper error record when builder subds isn't found
Previously, only an uniformative AssertionError would have been
reported. The error message now hints at the underlying cause for
the error (failure to find a builder subds) and hints at a likely
reason (command wasn't run in the superdataset, or the superdataset
did not yet receive a builder subds), and is an actual datalad
result record.
The old assertion statement is transformed into a second check
which is unlikely to ever error, but if it does, it would hint
at a general problem - either the subdataset result record
would have returned several items, or changed its reporting
structure. In those cases, the function raises a
RuntimeError to prevent downstream errors
Fixes #57.