Update distribution subdatasets of archive dataset before diff'ing changes
This makes sure that we get all relevant package change. The update is
done in a way that ensure we will only get package dataset updates that
have already been registered in their respective distributions, rather
than any updates of package datasets.

This is relevant, because the respective datasets might be maintained
differently, and a developed package dataset update might not yet be
approved for inclusion in a distribution.

Closes psychoinformatics-de/datalad-debian#102