Distinguish 'Debian release' and 'distribution' in doc
This commit modifies the walkthrough for distribution
dataset creation and for package building. It clearly
separates referencs tp DataLad's distribution datasets
and to Debian releases by refering to the former as
"distribution" or "distribution dataset", and to the
latter as "Debian release", or "Debian point release"
if point releases are referenced.

A few typos are corrected.