Faciliate build environment customizations
This change makes it possible to simply add files to drop into a build
environment to the builder dataset. This is a generic mechanism, which
also includes "finalizer" executables inside the environment, that can
be used for arbitrary customization -- including common ones, like
adding additional APT sources.

See docs of `deb-configure-builder` for more info.

Closes psychoinformatics-de/datalad-debian#33